pyproject.toml
[project]
name = "flawless"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "decord>=0.6.0",
    "ipykernel>=7.1.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "ultralytics>=8.3.222",
]

